requirements.txt
amqp==5.0.6
ansible==2.8.3
anyjson==0.3.3
appnope==0.1.0
asgiref==3.4.1
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.4.0
cachetools==4.2.4
celery==5.1.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
decorator==4.4.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-redis-cache==2.0.0
django-storages==1.12.3
django-timezone-field==4.2.1
django-widget-tweaks==1.4.5
djangorestframework==3.12.4
Faker==8.12.1
faker-microservice==2.0.0
google-api-core==2.7.1
google-auth==1.35.0
google-cloud-core==2.2.3
google-cloud-secret-manager==2.8.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
idna==2.8
ipip-ipdb==1.6.0
ipython-genutils==0.2.0
jedi==0.14.1
Jinja2==2.10.1
kombu==5.1.0
libcst==0.3.21
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==21.0
paramiko==2.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipdeptree==2.1.0
prompt-toolkit==2.0.9
proto-plus==1.19.2
protobuf==3.17.3
ptyprocess==0.6.0
pyasn1==0.4.6
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.4.1
redis==3.2.1
requests==2.22.0
rsa==4.7.2
six==1.16.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.25.3
uWSGI==2.0.18
vine==5.0.0
wcwidth==0.1.7
whitenoise==5.3.0
